+++ docs/changelog.md
@@ -17,6 +17,9 @@ Requires libvips v8.11.0
 * Allow multiple platform-arch binaries in same `node_modules` installation tree.
   [#2575](https://github.com/lovell/sharp/issues/2575)
 
+* Default to single-channel `b-w` space when `extractChannel` is used.
+  [#2658](https://github.com/lovell/sharp/issues/2658)
+
 ## v0.28 - *bijou*
 
 Requires libvips v8.10.6